+++ theme/template/functions/theme.php
@@ -1,0 +1,3 @@
+<?php
+// TODO: implement
+add_theme_support('post-thumbnails');