+++ theme/template/page.php
@@ -2,8 +2,12 @@
 get_header();
 
 switch($post->post_name) {
-  case 'news': {
+  case NEWS_CATEGORY: {
     include(get_template_directory().'/template-parts/news.php');
+  }
+
+  default: {
+    include(get_template_directory().'/template-parts/page-default.php');
   }
 }